refactor: ViewportPosition and CameraId move to the geometry leaf
Step 7 part 3 (the L-4 de-coupling): the pose currency and source
identity lived in the AI layer, forcing geometry to import detect. Both
relocate to reco_core::geometry (now public - these types are the
workspace-wide vocabulary) and every importer repoints. No re-export
from detect: the repo bans alias shims pre-1.0, so the roadmap's
one-release re-export is deliberately skipped in favor of fixing all
36 import sites outright.
