+++ crates/reco-core/src/session/tests.rs
@@ -12,11 +12,13 @@ use std::sync::{Arc, Mutex};
 use super::StitchSession;
 use super::types::*;
 use crate::calibration::{Calibration, Framing, Lens, Topology};
-use crate::detect::detector::{CameraId, Detection, DetectorError, DetectorFrame, UnifiedDetector};
-use crate::detect::director::{MappedDetection, ViewportPosition};
+use crate::detect::detector::{Detection, DetectorError, DetectorFrame, UnifiedDetector};
+use crate::detect::director::MappedDetection;
 use crate::detect::panner::{PanContext, Panner};
 use crate::detect::tracker::{TrackState, TrackedEntity, Tracker, WorldState};
 use crate::encoder::{EncodeError, Encoder, OutputFrame};
+use crate::geometry::CameraId;
+use crate::geometry::ViewportPosition;
 use crate::render::viewport::ViewportConfig;
 use crate::source::{FramePair, FrameSource, SourceError, SourceInfo, StereoFrame, YuvData};
