+++ crates/reco-detect/src/detectors/metal.rs
@@ -16,9 +16,8 @@ use std::path::Path;
 
 use crate::coreml_inference::CoreMlModel;
 use crate::metal_compute::MetalPreprocessPipeline;
-use reco_core::detect::detector::{
-    CameraId, Detection, DetectorError, DetectorFrame, UnifiedDetector,
-};
+use reco_core::detect::detector::{Detection, DetectorError, DetectorFrame, UnifiedDetector};
+use reco_core::geometry::CameraId;
 use reco_core::gpu::GpuContext;
 use reco_core::interop::metal::CVPixelBufferRef;
