+++ crates/reco-detect/src/detectors/ncnn.rs
@@ -20,8 +20,9 @@ use std::os::raw::c_char;
 use std::path::Path;
 
 use reco_core::detect::detector::{
-    CameraId, ChromaFormat, Detection, DetectorError, DetectorFrame, RawFrame, UnifiedDetector,
+    ChromaFormat, Detection, DetectorError, DetectorFrame, RawFrame, UnifiedDetector,
 };
+use reco_core::geometry::CameraId;
 
 // ── NCNN C API FFI ──────────────────────────────────────────────────
