+++ crates/reco-detect/src/detectors/trt/mod.rs
@@ -27,8 +27,9 @@ use crate::cuda_kernels::normalize_hwc_to_chw;
 use crate::npp_interop::{NppiRect, npp_mirror_c3, npp_nv12_to_rgb, npp_resize_c3};
 use reco_core::detect::detector::ChromaFormat;
 use reco_core::detect::detector::{
-    CameraId, Detection, DetectorError, DetectorFrame, GpuNv12Frame, UnifiedDetector,
+    Detection, DetectorError, DetectorFrame, GpuNv12Frame, UnifiedDetector,
 };
+use reco_core::geometry::CameraId;
 use reco_core::interop::cuda::{
     CUdeviceptr, cuda_ensure_context, cuda_mem_alloc, cuda_mem_free, cuda_memcpy_dtoh,
     cuda_memcpy_htod_2d, cuda_memset_d8, cuda_synchronize,
