+++ crates/reco-gui/src/main.rs
@@ -34,7 +34,7 @@ use reco_calibrate::{LensProfileInfo, ProfileSource};
 use reco_control::pose_control::{PoseControl, PoseControlConfig};
 use reco_control::{ControlIntent, PoseIntent};
 use reco_core::calibration::Calibration;
-use reco_core::detect::director::ViewportPosition;
+use reco_core::geometry::ViewportPosition;
 use reco_core::wgpu;
 
 use crate::playback::{PlayState, Playback};
